Josh C-S Files
It is an experimental program. It has nothing to do with sending/receiving data from other hosts yet.

All it does is reading a peer info from an external text file, and update the info according to the ones recorded in PeerInfo class.

In the test.txt file, you will see the contents of:
000.000.111.111
false
111.111.000.000
false
222.333.444.555
false

Upon running and completing the program, the content of the file will be changed to:
000.000.111.111
true
111.111.000.000
true
222.333.444.555
true

The sole purpose of this program is to have each host a list of host info up to date, so that it will keep an updated list of the participating hosts when it dies and returns to the protocol later.

Before test running this code, be sure to change the hard-coded directory to the directory where the test.txt file is located.

Also, the code is using an "absolute path" to access the external file. If you guys somehow know how to use a "relative path" to access the file, please share with us so that we can improve this code.